+++ src/LRMS.Application/Tables/Dto/TableDto.cs
@@ -5,4 +5,4 @@
 /// </summary>
 /// <param name="Id">Идентификатор.</param>
 /// <param name="SeatsCount">Количество мест.</param>
-public record struct TableDto(int Id, byte SeatsCount);
+public record struct TableDto(long Id, byte SeatsCount);
