+++ src/LRMS.Infrastructure/ReservationManagerApi/Books/Dto/CreateBookReservationCommand.cs
@@ -7,4 +7,4 @@
 /// <param name="Date">Дата бронирования.</param>
 /// <param name="CustomerName">Имя клиента.</param>
 /// <param name="CustomerPhone">Телефон клиента.</param>
-public record struct CreateBookReservationCommand(int BookId, DateTime Date, string CustomerName, string CustomerPhone);
+public record struct CreateBookReservationCommand(long BookId, DateTime Date, string CustomerName, string CustomerPhone);
