+++ src/LRMS.Infrastructure/ReservationManagerApi/Books/IBookApi.cs
@@ -9,9 +9,9 @@ public interface IBookApi
     Task<HttpResponseMessage> CreateBookReservation([Body] CreateBookReservationDto dto, CancellationToken ct = default);
 
     [Get("/api/v1/books")]
-    Task<HttpResponseMessage> GetBookReservations([Refit.Query] int? book_id, [Refit.Query] string? reserved_at,
+    Task<HttpResponseMessage> GetBookReservations([Refit.Query] long? book_id, [Refit.Query] string? reserved_at,
         [Refit.Query] int? page_number, [Refit.Query] int? page_size, CancellationToken ct = default);
 
     [Delete("/api/v1/books/{id}")]
-    Task<HttpResponseMessage> DeleteBookReservation(int id, CancellationToken ct = default);
+    Task<HttpResponseMessage> DeleteBookReservation(long id, CancellationToken ct = default);
 }
