+++ src/LRMS.Infrastructure/ReservationManagerApi/ReservationRequests/Dto/QuickReservationModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace LRMS.Infrastructure.ReservationManagerApi.ReservationRequests.Dto;
 
-public record struct QuickReservationModel(int id, string status, string name, string phone);
+public record struct QuickReservationModel(long id, string status, string name, string phone);
