+++ src/LRMS.Infrastructure/ReservationManagerApi/ReservationRequests/Dto/ReservationRequestDto.cs
@@ -7,4 +7,4 @@
 /// <param name="Status">Статус заявки на бронирование.</param>
 /// <param name="CustomerName">Имя клиента.</param>
 /// <param name="CustomerPhone">Номер телефона клиента.</param>
-public record struct ReservationRequestDto(int Id, string Status, string CustomerName, string CustomerPhone);
+public record struct ReservationRequestDto(long Id, string Status, string CustomerName, string CustomerPhone);
