+++ src/LRMS.Infrastructure/ReservationManagerApi/ReservationRequests/IReservationRequestRepository.cs
@@ -8,5 +8,5 @@ public interface IReservationRequestRepository
 {
     Task CreateReservationRequest(CreateReservationRequestCommand command, CancellationToken ct = default);
     Task<GetReservationRequestsResponse> GetReservationRequests(GetReservationRequest request, CancellationToken ct = default);
-    Task<ReservationRequestDto> UpdateReservationRequest(int id, UpdateReservationRequestCommand command, CancellationToken ct = default);
+    Task<ReservationRequestDto> UpdateReservationRequest(long id, UpdateReservationRequestCommand command, CancellationToken ct = default);
 }
