+++ src/LRMS.Infrastructure/ReservationManagerApi/Tables/Commands/CreateTableReservationCommand.cs
@@ -8,4 +8,4 @@
 /// <param name="CustomerPhone">Номер телефона клиента.</param>
 /// <param name="StartTime">Время начала.</param>
 /// <param name="EndTime">Время конца.</param>
-public record struct CreateTableReservationCommand(int TableId, string CustomerName, string CustomerPhone, DateTime StartTime, DateTime EndTime);
+public record struct CreateTableReservationCommand(long TableId, string CustomerName, string CustomerPhone, DateTime StartTime, DateTime EndTime);
