+++ src/LRMS.Infrastructure/ReservationManagerApi/Tables/Dto/TableReservationDto.cs
@@ -9,4 +9,4 @@
 /// <param name="CustomerPhone">Номер телефона клиента.</param>
 /// <param name="StartTime">Время начала.</param>
 /// <param name="EndTime">Время конца.</param>
-public record struct TableReservationDto(int Id, int TableId, string CustomerName, string CustomerPhone, DateTime StartTime, DateTime EndTime);
+public record struct TableReservationDto(long Id, long TableId, string CustomerName, string CustomerPhone, DateTime StartTime, DateTime EndTime);
