+++ src/LRMS.Infrastructure/ReservationManagerApi/Tables/Dto/TableReservationModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace LRMS.Infrastructure.ReservationManagerApi.Tables.Dto;
 
-public record struct TableReservationModel(int id, int table_id, string name, string phone, DateTime start_at, DateTime end_at);
+public record struct TableReservationModel(long id, long table_id, string name, string phone, DateTime start_at, DateTime end_at);
