+++ src/LRMS.Infrastructure/ReservationManagerApi/Tables/ITableReservationRepository.cs
@@ -6,9 +6,9 @@ namespace LRMS.Infrastructure.ReservationManagerApi.Tables;
 
 public interface ITableReservationRepository
 {
-    Task<IReadOnlyCollection<ReservationSlotDto>> GetSlots(int tableId, DateTime date, CancellationToken ct = default);
+    Task<IReadOnlyCollection<ReservationSlotDto>> GetSlots(long tableId, DateTime date, CancellationToken ct = default);
     Task CreateTableReservation(CreateTableReservationCommand tableReservationDto, CancellationToken ct = default);
-    Task<GetTableReservationsResponse> GetTableReservations(int? tableId, string? activeAt, int? pageNumber,
+    Task<GetTableReservationsResponse> GetTableReservations(long? tableId, string? activeAt, int? pageNumber,
         int? pageSize, CancellationToken ct = default);
-    Task DeleteTableReservation(int id, CancellationToken ct = default);
+    Task DeleteTableReservation(long id, CancellationToken ct = default);
 }
